executor: migrate test-infra to testify for executor_test.go testSlowQuerySuite
